+++ src/main/java/com/backend/phoneshop/dto/data/SaleProductMonthlyReport.java
@@ -16,7 +16,7 @@ public record SaleProductMonthlyReport(
         String brandName,
         Long categoryTypeId,
         String categoryTypeName,
-        Integer soleProductCount, // sum of quantity of sold product per month
+        BigDecimal soleProductCount, // sum of quantity of sold product per month
         BigDecimal totalPaymentAmount, // sum of paymentAmount per month
         BigDecimal totalPaidAmount // sum of paidAmount per month
 ) {
